Biome generation: Fix layers of 'filler' nodes at biome y limits
Error was exposed by commit 8cc9ce574bd80901e566a5838cc4504c20d897b8
'nplaced' was not set to U16_MAX when biome 'stone' was placed, so when
biome was recalculated in a column of stone, the 'nplaced' value
caused a few remaining filler nodes to be placed.
Occurs when the lower biome has a deeper depth of 'top' plus 'filler'
than the upper biome.